refactor the footer element insert
