+++ samples/image_to_text/index.html
@@ -313,16 +313,7 @@
       </div>
     </div>
 
-    <div class="divider"></div>
-
-    <footer>
-      <a
-        href="../../terms.html"
-        target="_blank"
-        class="text-stone-50/80 hover:text-stone-50 hover:underline hover:underline-offset-4 font-serif font-medium"
-        >Terms of Use</a
-      >
-    </footer>
+    <footer id="footer"></footer>
   </body>
 
   <script type="module" src="./image_to_text.js"></script>
