+++ samples/llm_gemma/gemma.html
@@ -373,16 +373,7 @@
       </div>
     </div>
 
-    <div class="divider"></div>
-
-    <footer>
-      <a
-        href="../../terms.html"
-        target="_blank"
-        class="text-stone-50/80 hover:text-stone-50 hover:underline hover:underline-offset-4 font-serif font-medium"
-        >Terms of Use</a
-      >
-    </footer>
+    <footer id="footer"></footer>
   </body>
 
   <script type="module" src="gemma.js"></script>
