+++ samples/question_answering/index.html
@@ -215,16 +215,7 @@ What is the drainage area of the Amazon rainforest?</textarea
       </section>
     </div>
 
-    <div class="divider"></div>
-
-    <footer>
-      <a
-        href="../../terms.html"
-        target="_blank"
-        class="text-stone-50/80 hover:text-stone-50 hover:underline hover:underline-offset-4 font-serif font-medium"
-        >Terms of Use</a
-      >
-    </footer>
+    <footer id="footer"></footer>
   </body>
 
   <script type="module" src="./question_answering.js"></script>
