+++ samples/stable_diffusion/stable-diffusion.html
@@ -537,16 +537,7 @@
       <div id="status" style="font: 1em consolas"></div>
     </div>
 
-    <div class="divider"></div>
-
-    <footer>
-      <a
-        href="../../terms.html"
-        target="_blank"
-        class="text-stone-50/80 hover:text-stone-50 hover:underline hover:underline-offset-4 font-serif font-medium"
-        >Terms of Use</a
-      >
-    </footer>
+    <footer id="footer"></footer>
   </body>
 
   <script type="module" src="./stable-diffusion.js"></script>
