+++ samples/summarization/index.html
@@ -295,16 +295,7 @@ The tower is 324 metres (1,063 ft) tall, about the same height as an 81-storey b
       </section>
     </div>
 
-    <div class="divider"></div>
-
-    <footer>
-      <a
-        href="../../terms.html"
-        target="_blank"
-        class="text-stone-50/80 hover:text-stone-50 hover:underline hover:underline-offset-4 font-serif font-medium"
-        >Terms of Use</a
-      >
-    </footer>
+    <footer id="footer"></footer>
   </body>
 
   <script type="module" src="./summarization.js"></script>
