+++ style/tailwind.css
@@ -945,6 +945,10 @@ video {
   -webkit-line-clamp: none;
 }
 
+.\!block {
+  display: block !important;
+}
+
 .block {
   display: block;
 }
@@ -1066,12 +1070,16 @@ video {
   height: 2rem;
 }
 
-.h-\[250px\] {
-  height: 250px;
+.h-\[245px\] {
+  height: 245px;
 }
 
-.h-\[256px\] {
-  height: 256px;
+.h-\[24px\] {
+  height: 24px;
+}
+
+.h-\[250px\] {
+  height: 250px;
 }
 
 .h-\[300px\] {
@@ -1098,6 +1106,10 @@ video {
   height: 512px;
 }
 
+.h-\[64px\] {
+  height: 64px;
+}
+
 .h-\[80px\] {
   height: 80px;
 }
@@ -1113,38 +1125,6 @@ video {
 
 .h-screen {
   height: 100vh;
-}
-
-.h-\[200px\] {
-  height: 200px;
-}
-
-.h-auto {
-  height: auto;
-}
-
-.h-\[100px\] {
-  height: 100px;
-}
-
-.h-\[72px\] {
-  height: 72px;
-}
-
-.h-\[64px\] {
-  height: 64px;
-}
-
-.h-\[24px\] {
-  height: 24px;
-}
-
-.h-\[260px\] {
-  height: 260px;
-}
-
-.h-\[245px\] {
-  height: 245px;
 }
 
 .max-h-\[200px\] {
@@ -1239,6 +1219,10 @@ video {
   width: 200px;
 }
 
+.w-\[24px\] {
+  width: 24px;
+}
+
 .w-\[450px\] {
   width: 450px;
 }
@@ -1299,14 +1283,6 @@ video {
 
 .w-screen {
   width: 100vw;
-}
-
-.w-3\/5 {
-  width: 60%;
-}
-
-.w-\[24px\] {
-  width: 24px;
 }
 
 .min-w-6 {
@@ -2750,10 +2726,6 @@ video {
   padding: 2rem;
 }
 
-.p-16 {
-  padding: 4rem;
-}
-
 .px-1 {
   padding-left: 0.25rem;
   padding-right: 0.25rem;
@@ -3128,8 +3100,8 @@ video {
   color: rgb(250 250 249 / var(--tw-text-opacity));
 }
 
-.text-stone-50\/50 {
-  color: rgb(250 250 249 / 0.5);
+.text-stone-50\/80 {
+  color: rgb(250 250 249 / 0.8);
 }
 
 .text-stone-500 {
@@ -3155,22 +3127,6 @@ video {
 .text-white {
   --tw-text-opacity: 1;
   color: rgb(255 255 255 / var(--tw-text-opacity));
-}
-
-.text-stone-50\/80 {
-  color: rgb(250 250 249 / 0.8);
-}
-
-.text-stone-50\/70 {
-  color: rgb(250 250 249 / 0.7);
-}
-
-.text-stone-50\/30 {
-  color: rgb(250 250 249 / 0.3);
-}
-
-.text-stone-50\/90 {
-  color: rgb(250 250 249 / 0.9);
 }
 
 .underline {
@@ -3970,10 +3926,6 @@ video {
     margin-top: 0.75rem;
   }
 
-  .\32xl\:mt-4 {
-    margin-top: 1rem;
-  }
-
   .\32xl\:mt-6 {
     margin-top: 1.5rem;
   }
@@ -4018,8 +3970,16 @@ video {
     height: 220px;
   }
 
+  .\32xl\:h-\[30px\] {
+    height: 30px;
+  }
+
   .\32xl\:h-\[360px\] {
     height: 360px;
+  }
+
+  .\32xl\:h-\[36px\] {
+    height: 36px;
   }
 
   .\32xl\:h-\[40px\] {
@@ -4032,22 +3992,6 @@ video {
 
   .\32xl\:h-\[512px\] {
     height: 512px;
-  }
-
-  .\32xl\:h-auto {
-    height: auto;
-  }
-
-  .\32xl\:h-\[200px\] {
-    height: 200px;
-  }
-
-  .\32xl\:h-\[36px\] {
-    height: 36px;
-  }
-
-  .\32xl\:h-\[30px\] {
-    height: 30px;
   }
 
   .\32xl\:min-h-\[28px\] {
@@ -4082,20 +4026,20 @@ video {
     width: 287px;
   }
 
-  .\32xl\:w-\[720px\] {
-    width: 720px;
-  }
-
-  .\32xl\:w-full {
-    width: 100%;
+  .\32xl\:w-\[30px\] {
+    width: 30px;
   }
 
   .\32xl\:w-\[36px\] {
     width: 36px;
   }
 
-  .\32xl\:w-\[30px\] {
-    width: 30px;
+  .\32xl\:w-\[720px\] {
+    width: 720px;
+  }
+
+  .\32xl\:w-full {
+    width: 100%;
   }
 
   .\32xl\:min-w-\[68px\] {
@@ -4160,8 +4104,6 @@ video {
     border-radius: 1.5rem;
   }
 
-  .\32xl\:p-4 {
-    padding: 1rem;
   .\32xl\:p-16 {
     padding: 4rem;
   }
