+++ terms.html
@@ -376,7 +376,7 @@
     </div>
 
     <script type="module">
-      import { setupNavigBar } from "./js/navbar.js";
+      import { setupNavigBar } from "./js/navigation.js";
 
       setupNavigBar(".");
     </script>
